moved zeek worker config generator to config module added equivalent method in suricata
